backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
PyMySQL==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
redis==5.0.1
gunicorn==21.2.0
waitress==3.0.2
